Move _postprocess_numbagg to aggregate_numbagg.py
- Relocates _postprocess_numbagg function from core.py to aggregate_numbagg.py
- Updates import in core.py to use the new location
- Groups numbagg-specific functionality together for better organization
